requirements.txt
functions-framework==3.*
requests
openai
python-dotenv
firebase-admin
configparser
tiktoken
langchain
faiss-cpu
SpeechRecognition==3.10.1
pyaudio==0.2.14
git+https://github.com/openai/whisper.git
sounddevice==0.4.6
obsws-python==1.7.0
watchdog==3.0.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client